backend/requirements.txt
flask
flask_sqlalchemy
flask_migrate
flask_jwt_extended
psycopg2-binary
python-dotenv
flask-bcrypt
flask-cors
